# docs/api-specs/gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Lumina API",
    "description": "D.S.A. Orchestrator + LLM Conversational Interface (multi-domain)",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/holodeck/simulate": {
      "post": {
        "summary": "Holodeck Simulate",
        "description": "Run a test message through the pipeline with proposed physics changes.",
        "operationId": "holodeck_simulate_api_holodeck_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HolodeckSimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolodeckSimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HolodeckSimulateRequest": {
        "properties": {
          "staged_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staged Id"
          },
          "physics_override": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physics Override"
          },
          "domain_id": {
            "type": "string",
            "title": "Domain Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "turn_data_override": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Data Override"
          },
          "deterministic_response": {
            "type": "boolean",
            "title": "Deterministic Response",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "domain_id",
          "message"
        ],
        "title": "HolodeckSimulateRequest",
        "description": "Run a test message through the pipeline with proposed physics changes.\n\nProvide *either* ``staged_id`` (referencing a pending HITL staged command\nwhose operation is ``update_domain_physics``) *or* ``physics_override``\n(an inline dict of physics fields to merge onto the live physics)."
      },
      "HolodeckSimulateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "response": {
            "type": "string",
            "title": "Response"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "prompt_type": {
            "type": "string",
            "title": "Prompt Type"
          },
          "escalated": {
            "type": "boolean",
            "title": "Escalated"
          },
          "tool_results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Results"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "structured_content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Content"
          },
          "sandbox_physics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Physics"
          },
          "physics_diff": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physics Diff"
          },
          "live_physics_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Physics Hash"
          },
          "sandbox_physics_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Physics Hash"
          },
          "staged_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staged Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "response",
          "action",
          "prompt_type",
          "escalated"
        ],
        "title": "HolodeckSimulateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}